Fix depth issues
This fixes the script to accurately track the depth of each link found on the page. If we had 4 links on the first page, with each link having 4 unique links and a depth set at 3, it should be able to crawl all pages. Instead, the script is fetching all links on the first page (depth 1), the links from the first link (depth 2), and will ignore the links from link 2 - 4 (since they are all first seen as depth 3 from the first link).